Add a test for form:errors tags
